pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "local-coding-agent"
version = "0.1.0"
description = "A local coding agent with LLM integration, MCP tools, and multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Developer", email = "dev@example.com" }
]
keywords = ["coding-agent", "llm", "mcp", "autonomous-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=1.1.0",
    "mcp>=1.0.0",
    "chromadb>=0.6.0",
    "networkx>=3.3",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.20.0",
    "tenacity>=8.2.0",
    "structlog>=24.2.0",
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.4.0",
    "isort>=5.13.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
ferrite = "local_coding_agent.__main__:main"

[tool.poetry]
name = "local-coding-agent"
version = "0.1.0"
description = "A local coding agent with LLM integration, MCP tools, and multi-agent orchestration"
authors = ["Developer <dev@example.com>"]
packages = [{ include = "local_coding_agent" }]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^1.1"
mcp = "^1.0"
chromadb = "^0.6"
networkx = "^3.3"
pydantic = "^2.8"
pydantic-settings = "^2.3"
sqlalchemy = "^2.0"
python-dotenv = "^1.0"
prometheus-client = "^0.20"
tenacity = "^8.2"
structlog = "^24.2"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
black = "^24.4"
isort = "^5.13"
mypy = "^1.10"
ruff = "^0.4"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
